// frontend/web/src/services/API/proserve-wb-provisioning-api/models/LaunchProductRequest.ts
export interface LaunchProductRequest {
    /**
     * Product ID.
     * @type {string}
     * @memberof LaunchProductRequest
     */
    productId: string;
    /**
     * Unique ID of the product version to be provisioned.
     * @type {string}
     * @memberof LaunchProductRequest
     */
    versionId: string;
    /**
     * Parameters specified by the administrator that are required for provisioning the product.
     * @type {Array<ProvisioningParameter>}
     * @memberof LaunchProductRequest
     */
    provisioningParameters?: Array<ProvisioningParameter>;
    /**
     * Additional configurations that are required to configure the provisioned product.
     * @type {Array<AdditionalConfiguration>}
     * @memberof LaunchProductRequest
     */
    additionalConfigurations?: Array<AdditionalConfiguration>;
    /**
     * Product provisioning stage.
     * @type {string}
     * @memberof LaunchProductRequest
     */
    stage: string;
    /**
     * Region to which provision the product.
     * @type {string}
     * @memberof LaunchProductRequest
     */
    region: string;
}